\name{bamboo.MSA.astral30}
\docType{data}
\alias{bamboo.MSA.astral30}
\title{Multiple sequence alignment (MSA) information for the bamboo.validation.astral30 test dataset in the paper "Bayesian Model of Protein Primary Sequence for Secondary Structure Prediction"} 
\description{
This MSA information list gives the count matrices for the 3,143 proteins in the bamboo.validation.astral30 test dataset that have MSA information. Each row in the matrix is the count vector for the number of times that each of the four secondary structure types is found in that position of the alignment output.}
\usage{bamboo.MSA.astral30}
\format{A list containing 3,143 matrices.}
\source{Li, Dahl, Vannucci, Joo, Tsai (2014), Bayesian Model of Protein Primary Sequence for Secondary Structure Prediction}
\references{\url{http://dahl.byu.edu/software/bamboo/}}
\keyword{datasets}
