\name{bamboo.MSA.casp9}
\docType{data}
\alias{bamboo.MSA.casp9}
\title{Multiple sequence alignment (MSA) information for the bamboo.validation.casp9 test dataset in the paper "Bayesian Model of Protein Primary Sequence for Secondary Structure Prediction"} 
\description{
This MSA information list gives the count matrices for the 109 proteins in the bamboo.validation.casp9 test dataset that have MSA information. Each row in the matrix is the count vector for the number of times that each of the four secondary structure types is found in that position of the alignment output.}
\usage{bamboo.MSA.casp9}
\format{A list containing 109 matrices.}
\source{Li, Dahl, Vannucci, Joo, Tsai (2014), Bayesian Model of Protein Primary Sequence for Secondary Structure Prediction}
\references{\url{http://dahl.byu.edu/software/bamboo/}}
\keyword{datasets}
