\name{bamboo.validation.astral30}
\docType{data}
\alias{bamboo.validation.astral30}
\title{Validation (test) dataset named astral30 in the paper "Bayesian Model of Protein Primary Sequence for Secondary Structure Prediction"} 
\description{
This validation dataset gives the names, the primary structure (amino acid sequences), and the secondary structure of 3,344 individual proteins from the SCOPe 2.03 data set filtered at 30\% sequence identity.}
\usage{bamboo.validation.astral30}
\format{A data frame containing 3,344 observations on the following 3 variables.
\enumerate{
\item name: protein name;
\item primary: protein primary structure (amino acid sequence) in 20 letters denoting the 20 amino acids;
\item hetc: secondary structure in 4 letters denoting the 4 structure types: helix (H), strand (E), turn (T) and coil (C).
}}
\source{Fox NK, Brenner SE, Chandonia JM (2013) Scope: Structural classification of proteins extended, integrating scop and astral data and classification of new structures. \emph{Nucleic Acids Research} 42: D304-D309}
\references{\url{http://dahl.byu.edu/software/bamboo/}}
\keyword{datasets}
