\name{bamboo.validation.casp9}
\docType{data}
\alias{bamboo.validation.casp9}
\title{Validation (Test) Dataset Named casp9} 
\description{
This validation dataset gives the names, the primary structure (amino acid sequences), and the secondary structure of 203 individual proteins from the targets used in CASP9 experiments as used in the paper cited below.}
\usage{bamboo.validation.casp9}
\format{A data frame containing 203 observations on the following 3 variables.
\enumerate{
\item name: protein name;
\item primary: protein primary structure (amino acid sequence) in 20 letters denoting the 20 amino acids;
\item hetc: secondary structure in 4 letters denoting the 4 structure types: helix (h), strand (e), turn (t) and coil (c).
}}
\source{Moult J, Fidelis K, Kryshtafovych A, Tramontano A (2011) Critical assessment of methods of protein structure prediction (casp) round ix. \emph{Proteins: Structure, Function, and Bioinformatics} 79: 1-5. <DOI:10.1002/prot.23200>}
\references{Q. Li, D. B. Dahl, M. Vannucci, H. Joo, J. W. Tsai (2014), Bayesian Model of Protein Primary Sequence for Secondary Structure Prediction, PLOS ONE, 9(10), e109832. <DOI:10.1371/journal.pone.0109832>}
\keyword{datasets}
