% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_panel.R
\name{check_panel}
\alias{check_panel}
\title{Check if Data Has Panel Attributes}
\usage{
check_panel(data)
}
\arguments{
\item{data}{A data.frame to check.}
}
\value{
Logical value indicating whether the data has panel attributes.
}
\description{
Helper function to check if a data.frame has panel attributes.
}
\examples{
data(production)
check_panel(production)  # FALSE

panel_data <- set_panel(production, group = "firm", time = "year")
check_panel(panel_data)  # TRUE

}
