% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_if_needed}
\alias{round_if_needed}
\title{Round numeric values if needed}
\usage{
round_if_needed(x, digits)
}
\arguments{
\item{x}{A vector (typically numeric).}

\item{digits}{Number of decimal places.}
}
\value{
Rounded vector if numeric and not all NA, otherwise unchanged.
}
\description{
Round numeric values if needed
}
\keyword{internal}
