% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_unique_preserve}
\alias{sort_unique_preserve}
\title{Sort unique values preserving original type where sensible}
\usage{
sort_unique_preserve(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Sorted unique values.
}
\description{
Sort unique values preserving original type where sensible
}
\keyword{internal}
