\name{bamboo.MSA.astral30}
\docType{data}
\alias{bamboo.MSA.astral30}
\title{MSA Information for the bamboo.validation.astral30 Test Dataset} 
\description{
This data provides the multiple sequence alignment (MSA) information for the \code{bamboo.validation.astral30} test dataset in the paper cited below.
This MSA information list gives the count matrices for the 3,143 proteins in the bamboo.validation.astral30 test dataset that have MSA information. Each row in the matrix is the count vector for the number of times that each of the four secondary structure types is found in that position of the alignment output.}
\usage{bamboo.MSA.astral30}
\format{A list containing 3,143 matrices.}
\source{Q. Li, D. B. Dahl, M. Vannucci, H. Joo, J. W. Tsai (2014), Bayesian Model of Protein Primary Sequence for Secondary Structure Prediction, PLOS ONE, 9(10), e109832. <DOI:10.1371/journal.pone.0109832>}
\references{Q. Li, D. B. Dahl, M. Vannucci, H. Joo, J. W. Tsai (2014), Bayesian Model of Protein Primary Sequence for Secondary Structure Prediction, PLOS ONE, 9(10), e109832. <DOI:10.1371/journal.pone.0109832>}
\keyword{datasets}
