\name{bamboo.MSA.casp9}
\docType{data}
\alias{bamboo.MSA.casp9}
\title{MSA Information for the bamboo.validation.casp9 Test Dataset} 
\description{
This data provides the multiple sequence alignment (MSA) information for the \code{bamboo.validation.casp9} test dataset in the paper cited below.
This MSA information list gives the count matrices for the 109 proteins in the bamboo.validation.casp9 test dataset that have MSA information. Each row in the matrix is the count vector for the number of times that each of the four secondary structure types is found in that position of the alignment output.}
\usage{bamboo.MSA.casp9}
\format{A list containing 109 matrices.}
\references{
Q. Li, D. B. Dahl, M. Vannucci, H. Joo, J. W. Tsai (2014), Bayesian Model of Protein Primary Sequence for Secondary Structure Prediction, PLOS ONE, 9(10), e109832.
}
\keyword{datasets}
