% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_panel.R
\name{describe_panel}
\alias{describe_panel}
\title{Panel Data Description}
\usage{
describe_panel(data, group, time)
}
\arguments{
\item{data}{A data.frame containing panel data.}

\item{group}{A character string specifying the name of the entity/group variable in panel data.}

\item{time}{A character string specifying the name of the time variable.}
}
\value{
Invisible list containing panel description statistics.
}
\description{
Provides basic summary statistics for panel data structure.
}
\examples{
data(production)
describe_panel(production, group = "firm", time = "year")

}
